# lib/api-zod/src/generated/types/botSettingsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotSettingsResponse {
  success: boolean;
  systemPrompt: string;
  autoReplyEnabled: boolean;
}