# lib/api-zod/src/generated/types/botStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BotStatusStatus } from "./botStatusStatus";

export interface BotStatus {
  status: BotStatusStatus;
  error: string | null;
  startedAt: string | null;
  messagesHandled: number;
  autoReplyEnabled: boolean;
  systemPrompt: string;
}